Fix conversion compiler warnings
Using a release build with GCC 14.2.0 and -Wextra -Wconversion -Wall.
The generated binaries are not changed by this commit.

+++ taglib/asf/asfattribute.cpp
@@ -302,16 +302,16 @@ ByteVector ASF::Attribute::render(const String &name, int kind) const
 
   if(kind == 0) {
     data = renderString(name, true) +
-           ByteVector::fromShort(static_cast<int>(d->type), false) +
-           ByteVector::fromShort(data.size(), false) +
+           ByteVector::fromShort(static_cast<short>(d->type), false) +
+           ByteVector::fromShort(static_cast<short>(data.size()), false) +
            data;
   }
   else {
     ByteVector nameData = renderString(name);
-    data = ByteVector::fromShort(kind == 2 ? d->language : 0, false) +
-           ByteVector::fromShort(d->stream, false) +
-           ByteVector::fromShort(nameData.size(), false) +
-           ByteVector::fromShort(static_cast<int>(d->type), false) +
+    data = ByteVector::fromShort(static_cast<short>(kind == 2 ? d->language : 0), false) +
+           ByteVector::fromShort(static_cast<short>(d->stream), false) +
+           ByteVector::fromShort(static_cast<short>(nameData.size()), false) +
+           ByteVector::fromShort(static_cast<short>(d->type), false) +
            ByteVector::fromUInt(data.size(), false) +
            nameData +
            data;
